+++ src/qibo/quantum_info/__init__.py
@@ -3,7 +3,7 @@ from qibo.quantum_info.clifford import *
 from qibo.quantum_info.entanglement import *
 from qibo.quantum_info.entropies import *
 from qibo.quantum_info.metrics import *
-from qibo.quantum_info.operations import *
+from qibo.quantum_info.linalg_operations import *
 from qibo.quantum_info.quantum_networks import *
 from qibo.quantum_info.random_ensembles import *
 from qibo.quantum_info.superoperator_transformations import *
