+++ src/qibo/quantum_info/entanglement.py
@@ -5,7 +5,7 @@ import numpy as np
 from qibo.backends import _check_backend
 from qibo.config import PRECISION_TOL, raise_error
 from qibo.quantum_info.metrics import fidelity, purity
-from qibo.quantum_info.operations import partial_trace
+from qibo.quantum_info.linalg_operations import partial_trace
 
 
 def concurrence(state, bipartition, check_purity: bool = True, backend=None):
