+++ src/qibo/quantum_info/entropies.py
@@ -7,8 +7,8 @@ from scipy.linalg import fractional_matrix_power
 
 from qibo.backends import _check_backend
 from qibo.config import PRECISION_TOL, raise_error
+from qibo.quantum_info.linalg_operations import partial_trace
 from qibo.quantum_info.metrics import _check_hermitian_or_not_gpu, purity
-from qibo.quantum_info.operations import partial_trace
 
 
 def shannon_entropy(prob_dist, base: float = 2, backend=None):
