+++ src/qibo/quantum_info/linalg_operations.py
